iot_data_processor/requirements.txt
paho-mqtt
pydantic==1.*
boto3
